Implemented check for crawler to update models only if there was no error.

+++ app/Http/Controllers/CrawlerController.php
@@ -16,27 +16,29 @@ class CrawlerController extends Controller
         $crawledUrls = $request->json('crawledUrls');
         $mailServerDomainList = $request->json('mailServerDomainList');
 
-        $domain->crawledUrls()->update(['domain_id' => null]);
-        $domain->mailDomains()->detach();
+        if ($request->json('hasError') === false) {
+            $domain->crawledUrls()->update(['domain_id' => null]);
+            $domain->mailDomains()->detach();
 
-        if ($mainUrl) {
-            $domain->crawledUrls()->firstOrCreate([
-                'url' => $mainUrl,
-                'is_main_url' => true
-            ]);
-        }
-
-        if ($crawledUrls) {
-            foreach ($crawledUrls as $url) {
-                $crawledUrl = CrawledUrl::whereUrl($url)->firstOrCreate(['url' => $url]);
-                $crawledUrl->domain()->associate($domain)->save();
+            if ($mainUrl) {
+                $domain->crawledUrls()->firstOrCreate([
+                    'url' => $mainUrl,
+                    'is_main_url' => true
+                ]);
             }
-        }
 
-        if ($mailServerDomainList) {
-            foreach ($mailServerDomainList as $mailServerEntry) {
-                $mailDomain = MailDomain::whereDomain($mailServerEntry)->firstOrCreate(['domain' => $mailServerEntry]);
-                $domain->mailDomains()->attach($mailDomain);
+            if ($crawledUrls) {
+                foreach ($crawledUrls as $url) {
+                    $crawledUrl = CrawledUrl::whereUrl($url)->firstOrCreate(['url' => $url]);
+                    $crawledUrl->domain()->associate($domain)->save();
+                }
+            }
+
+            if ($mailServerDomainList) {
+                foreach ($mailServerDomainList as $mailServerEntry) {
+                    $mailDomain = MailDomain::whereDomain($mailServerEntry)->firstOrCreate(['domain' => $mailServerEntry]);
+                    $domain->mailDomains()->attach($mailDomain);
+                }
             }
         }
     }
